chore: move getAccessTokenForConnection and getCredentials
