--- packages/ai-langchain/src/FederatedConnections/getAccessTokenForConnection.ts
@@ -1,11 +1,0 @@
-import { asyncLocalStorage } from "@auth0/ai/FederatedConnections";
-
-export const getAccessTokenForConnection = () => {
-  const store = asyncLocalStorage.getStore();
-  if (typeof store === "undefined") {
-    throw new Error(
-      "The tool must be wrapped with the withFederatedConnections function."
-    );
-  }
-  return store?.accessToken;
-};
