+++ packages/ai-langchain/src/FederatedConnections/index.ts
@@ -1,2 +1,2 @@
 export { FederatedConnectionAuthorizer } from "./FederatedConnectionAuthorizer";
-export { getAccessTokenForConnection } from "./getAccessTokenForConnection";
+export { getAccessTokenForConnection } from "@auth0/ai/FederatedConnections";
