+++ packages/ai-vercel/src/CIBA/index.ts
@@ -1,2 +1,2 @@
 export { CIBAAuthorizer } from "./CIBAAuthorizer";
-export { getCIBACredentials } from "./getCIBACredentials";
+export { getCIBACredentials } from "@auth0/ai/CIBA";
