--- packages/ai-vercel/src/FederatedConnections/getAccessTokenForConnection.ts
@@ -1,11 +1,0 @@
-import { asyncLocalStorage } from "@auth0/ai/FederatedConnections";
-
-export const getAccessTokenForConnection = () => {
-  const t = asyncLocalStorage.getStore();
-  if (typeof t === "undefined") {
-    throw new Error(
-      "The tool must be wrapped with the withTokenForConnection function."
-    );
-  }
-  return t.accessToken;
-};
